# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Contact API with Authentication, Roles, and Permissions"
  },
  "basePath": "/api",
...
  "securityDefinitions": {
    "cookieAuth": {
      "type": "apiKey",
      "name": "session",
      "in": "cookie",
      "description": "Session cookie obtained after successful login. Handled automatically by the browser."
    }
  },
  "paths": {
...
    "/app/roles/{role_id}/permissions/{permission_id}": {
      "delete": {
        "summary": "Remove a permission from a role.",
        "tags": ["Role Management"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "The ID of the role."
          },
          {
            "name": "permission_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "The ID of the permission to remove."
          }
        ],
        "responses": {
          "200": {
            "description": "Permission removed successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "role": { "$ref": "#/definitions/Role" }
              }
            }
          },
          "401": { "$ref": "#/responses/UnauthorizedError" },
          "403": { "$ref": "#/responses/ForbiddenError" },
          "404": { "$ref": "#/responses/NotFoundError" }
        }
      }
    },
...
    "/app/permissions/{permission_id}": {
      "get": {
        "summary": "Get a single permission by ID.",
        "tags": ["Permission Management"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "The ID of the permission to retrieve."
          },
          {
            "name": "include_usage",
            "in": "query",
            "type": "boolean",
            "description": "If true, includes usage statistics (number of roles and affected role names) for the permission.",
            "default": false
          },
          {
            "name": "include_category_details",
            "in": "query",
            "type": "boolean",
            "description": "If true, includes the full category object for the permission.",
            "default": false
          }
        ],
        "responses": {
          "200": {
            "description": "Permission details.",
            "schema": { "$ref": "#/definitions/Permission" }
          },
          "401": { "$ref": "#/responses/UnauthorizedError" },
          "403": { "$ref": "#/responses/ForbiddenError" },
          "404": { "$ref": "#/responses/NotFoundError" }
        }
      },
      "patch": {
        "summary": "Update an existing permission.",
        "tags": ["Permission Management"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "The ID of the permission to update."
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/PermissionUpdateInput" }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission updated successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "permission": { "$ref": "#/definitions/Permission" }
              }
            }
          },
          "400": { "$ref": "#/responses/BadRequestError" },
          "401": { "$ref": "#/responses/UnauthorizedError" },
          "403": { "$ref": "#/responses/ForbiddenError" },
          "404": { "$ref": "#/responses/NotFoundError" },
          "409": { "$ref": "#/responses/ConflictError" }
        }
      },
      "delete": {
        "summary": "Delete a permission.",
        "tags": ["Permission Management"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "The ID of the permission to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Permission deleted successfully.",
            "schema": { "$ref": "#/definitions/SuccessMessage" }
          },
          "401": { "$ref": "#/responses/UnauthorizedError" },
          "403": { "$ref": "#/responses/ForbiddenError" },
          "404": { "$ref": "#/responses/NotFoundError" },
          "409": { "$ref": "#/responses/ConflictError" }
        }
      }
    },
...
  },
  "definitions": {
...
    "Role": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "description": { "type": "string", "x-nullable": true },
        "permissions": {
          "type": "array",
          "description": "List of permissions associated with this role.",
          "items": { "$ref": "#/definitions/Permission" }
        }
      }
    },
...
    "Permission": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "description": { "type": "string", "x-nullable": true },
        "category_id": { "type": "integer", "description": "The ID of the category the permission belongs to.", "x-nullable": true },
        "category": {
          "$ref": "#/definitions/CategoryBasic",
          "description": "The detailed category object for the permission. Only included if `include_category_details=true`.",
          "x-nullable": true
        },
        "status": { "type": "string", "enum": ["active", "inactive"], "description": "The active status of the permission.", "default": "active" },
        "usage": { "type": "integer", "description": "The number of roles this permission is currently assigned to. Only included if `include_usage=true`." },
        "affected_roles": {
          "type": "array",
          "description": "A list of roles (ID and Name) that this permission is assigned to. Only included if `include_usage=true`.",
          "items": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" }
            }
          }
        }
      }
    },
...
    "PermissionUpdateInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "The new unique name of the permission."
        },
        "description": {
          "type": "string",
          "description": "An updated description of the permission.",
          "x-nullable": true
        },
        "category_id": {
          "type": "integer",
          "description": "An updated category ID for the permission. Set to `null` to remove category association.",
          "x-nullable": true
        },
        "status": {
          "type": "string",
          "enum": ["active", "inactive"],
          "description": "The updated status of the permission."
        }
      }
    },
    "CategoryBasic": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "description": { "type": "string", "x-nullable": true }
      },
      "description": "Basic details of a category without usage information."
    },
    "SuccessMessage": {
      "type": "object",
      "properties": {
        "message": { "type": "string" }
      }
    },
    "ErrorMessage": {
      "type": "object",
      "properties": {
        "message": { "type": "string" }
      }
    },
...
  },
  "responses": {
    "UnauthorizedError": {
      "description": "Authentication information is missing or invalid. User is not logged in.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    },
    "ForbiddenError": {
      "description": "The user does not have the necessary permissions for a resource.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    },
    "NotFoundError": {
      "description": "The requested resource was not found.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    },
    "BadRequestError": {
      "description": "Invalid input provided.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    },
    "ConflictError": {
      "description": "Resource already exists (e.g., username, email, role name, permission name) or a dependency conflict.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    }
  }
}